Support streams.
This patch adds support for non-seekable devices. It also cleans the
code and simplifies the implementation.

1. Removed any actions from the destructor. A finish member-function
must be called.
2. Removed an overload for a container, there is no need to bloat the
interface.
3. Header is initialized at the creation time.
4. Removed unnecessary instance variables and member functions.

The seek function is no longer called without a purpose. At the finish
procedure, if the device is seekable, we're storing the table of
contents and update an offset in the header. Otherwise, we just skip
it.